feat: Implement onboarding screen UI and logic
- Add `OnboardingDataModel` for onboarding content.
- Create `CustomAnimatedWidget` for onboarding animations.
- Implement `OnboardingView` with page indicator and navigation.
- Update color and text styles for onboarding screen.
- Integrate `CustomMaterialButton` for navigation actions.
