+++ lib/core/theming/colors_manager.dart
@@ -3,23 +3,9 @@ import 'dart:ui';
 abstract class ColorsManager {
   static const white = Color(0xffFFFFFF);
   static const black = Color(0xff000000);
-  static const color5F33E1 = Color(0xff5F33E1);
-  static const colorE6E6E6 = Color(0xffE6E6E6);
-  static const color744EE5 = Color(0xff744EE5);
-  static const color404147 = Color(0xff404147);
-  static const colorAFAFAF = Color(0xffAFAFAF);
-  static const color817D8D = Color(0xff817D8D);
-  static const colorBABABA = Color(0xffBABABA);
-  static const color7F7F7F = Color(0xff7F7F7F);
-  static const color6E6A7C = Color(0xff6E6A7C);
-  static const colorFF4949 = Color(0xffFF4949);
-  static const googleSignInBorderColor = Color(0xffdadce0);
-  static const googleSignInTextColor = Color(0xff3c4043);
-  static const color24252C = Color(0xff24252C);
-  static const color272727 = Color(0xff272727);
-  static const color424242 = Color(0xff424242);
-  static const color979797 = Color(0xff979797);
-  static const colorE1E0E3 = Color(0xffE1E0E3);
-  static const resetButtonColor = Color(0xffF5F5F5);
   static const colorEBEBEB = Color(0xffEBEBEB);
+  static const colorAAA198 = Color(0xffAAA198);
+  static const color212121 = Color(0xff212121);
+  static const color1F1F1F = Color(0xff1F1F1F);
+  static const color5C5C5C = Color(0xff5C5C5C);
 }
