+++ lib/feature/onboarding/presentation/widgets/custom_animated_widget.dart
@@ -1,0 +1,28 @@
+import 'package:animate_do/animate_do.dart';
+import 'package:flutter/material.dart';
+
+class CustomAnimatedWidget extends StatelessWidget {
+  const CustomAnimatedWidget({
+    super.key,
+    required this.index,
+    required this.delay,
+    required this.child,
+  });
+  final int index;
+  final int delay;
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    if (index == 1) {
+      return FadeInDown(
+        delay: Duration(milliseconds: delay),
+        child: child,
+      );
+    }
+    return FadeInUp(
+      delay: Duration(milliseconds: delay),
+      child: child,
+    );
+  }
+}
